azure_requirements.txt
# Update requirements to include Azure Storage SDK
azure-storage-blob>=12.8.0
